requirements.txt
Django==5.0.2
python-dotenv==1.0.1
# psycopg2-binary==2.9.9
django-storages==1.14.2
boto3==1.34.34
django-debug-toolbar==4.3.0
django-filter==23.5
python-decouple==3.8
setuptools==69.2.0 
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.43.0
cryptography==3.4.8
dj-database-url==2.3.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-phonenumber-field==8.0.0
django-resized==1.0.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.4.1
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mysql==0.0.3
mysqlclient==2.2.7
packaging==24.2
phonenumbers==8.13.54
pillow==11.1.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
channels